// LLD/Design Patterns/ProxyPattern/src/Enums.ts
// Enums for roles and access types
export enum UserRole {
    EMPLOYEE = "EMPLOYEE",
    MANAGER = "MANAGER",
    ADMIN = "ADMIN",
    HR = "HR"
  }
  
  export enum AccessType {
    CREATE = "CREATE",
    DELETE = "DELETE",
    UPDATE = "UPDATE",
    VIEW_ALL_DETAILS = "VIEW_ALL_DETAILS",
    VIEW_BASIC_INFO = "VIEW_BASIC_INFO"
  }